Cleanup main so platform can import cmd.Run
