+++ client/cli/cli.go
@@ -11,6 +11,11 @@ import (
 	"github.com/micro/micro/v3/client/cli/util"
 	"github.com/micro/micro/v3/cmd"
 	"github.com/urfave/cli/v2"
+
+	_ "github.com/micro/micro/v3/client/cli/init"
+	_ "github.com/micro/micro/v3/client/cli/new"
+	_ "github.com/micro/micro/v3/client/cli/signup"
+	_ "github.com/micro/micro/v3/client/cli/user"
 )
 
 var (
