+++ cmd/platform/main.go
@@ -4,7 +4,24 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+
+	"github.com/micro/go-micro/v3/errors"
+	"github.com/micro/micro/v3/cmd"
+
+	// internal packages
+	_ "github.com/micro/micro/v3/internal/usage"
+
+	// load packages so they can register commands
+	_ "github.com/micro/micro/v3/client/cli"
+	_ "github.com/micro/micro/v3/server"
+	_ "github.com/micro/micro/v3/service/cli"
+
+	// include the platform profile
+	_ "github.com/micro/micro/v3/profile/platform"
 )
+
+func main() {
+}
 
 var (
 	image = "micro/platform"
@@ -55,7 +72,8 @@ func main() {
 			return
 		}
 	default:
-		fmt.Println(usage)
+		// run micro by default
+		cmd.Run()
 		return
 	}
 }
